+++ lib/ghost_autocomplete.dart
@@ -3,3 +3,5 @@ library ghost_autocomplete;
 export 'src/ghost_autocomplete_widget.dart';
 export 'src/ghost_autocomplete_controller.dart';
 export 'src/ghost_autocomplete_list_widget.dart';
+export 'src/ghost_spell_check_service.dart';
+export 'src/ghost_spell_check_text_field.dart';